pyproject.toml
[tool.poetry]
name = "referral-system-auth-profile-invite-codes"
version = "0.1.0"
description = ""
authors = ["md <marceldiomare@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
asyncpg = "^0.30.0"
load-dotenv = "^0.1.0"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
redis = "^5.2.0"
djangorestframework-simplejwt = "^5.3.1"
requests = "^2.32.3"
pytest = "^7.0"
pytest-django = "^4.5"
pytest-factoryboy = "^2.1"
drf-spectacular = "^0.28.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^6.0.0"

[tool.poetry.group.dev.dependencies]
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
